requirements.txt
Django==1.4.3
Fabric==1.5.2
MySQL-python==1.2.4
South==0.7.6
boto==2.7.0
dj-database-url==0.2.1
django-bootstrap-static-files==2.0.4
django-clean==0.1.2
django-debug-toolbar==0.9.4
django-storages==1.1.4
factory-boy==1.2.0
gunicorn==0.17.2
icalendar==3.2
ipdb==0.7
ipython==0.13.1
mailsnake==1.6.1
paramiko==1.9.0
pycrypto==2.6
pytz==2012j
requests==1.1.0
simplejson==3.0.7
ssh==1.7.14
tldextract==1.1
wsgiref==0.1.2
